refactor(review): honor core type module boundary
